fix: 로또가이드, 쪼개고 이름 변경
